Rename module and simplify API
