+++ app/main.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, FastAPI
 
 from app.v1.lifecycle.router import v1_router as lifecycle_v1_router
-from app.v1.released.endpoints import v1_router as released_v1_router
+from app.v1.release_notes.endpoints import v1_router as release_notes_v1_router
 from app.v1.upcoming.endpoints import v1_router as upcoming_v1_router
 
 # Initialize FastAPI app
@@ -11,7 +11,7 @@ app = FastAPI()
 api_router = APIRouter()
 
 # Include individual service routers under the main API router
-api_router.include_router(released_v1_router, prefix="/v1/release-notes", tags=["release-notes"])
+api_router.include_router(release_notes_v1_router, prefix="/v1/release-notes", tags=["release-notes"])
 api_router.include_router(upcoming_v1_router, prefix="/v1/upcoming-changes", tags=["upcoming-changes"])
 api_router.include_router(lifecycle_v1_router, prefix="/v1/lifecycle")
